Mark instances correctly as creatable or not correctly. (#81)
This prevents erroneously creating services using Instance.new, but
required a comprehensive review of every test that touches instances.

This establishes a new ad-hoc convention: for any instances that are
creatable, create them using Instance.new and include a test that does
just that. For instances that are not, import their module directly and
use Foo:new().